requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
commonmark==0.9.1
cookiecutter==2.2.3
cryptography==41.0.3
dataclasses-json==0.5.14
defusedxml==0.7.1
Django==4.1.11
django-allauth==0.55.2
django-bootstrap-v5==1.0.11
django-grappelli==3.0.7
django-tailwind==3.6.0
frozenlist==1.4.0
git-filter-repo==2.38.0
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
langchain==0.0.262
langsmith==0.0.19
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
text-unidecode==1.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3
yarl==1.9.2
whitenoise==6.5.0
azure-storage-blob==12.17.0
django-storages==1.14
daphne==4.0.0
channels==4.0.0
channels_redis==4.1.0
djangorestframework==3.14.0
django-filter==23.3.0
pypandoc==1.11.0
weaviate-client==3.24.1
llama-index==0.8.27